refactoring: distinguish Query and Response better
Make it a bit clearer which parts are in the general domain of running a
SPARQL query, and which are the business-logic of this specific query.

I suspect these will need to be separated out even further later as we
add additional queries, but this should hopefully make that a bit
simpler.