Don't warn about date errors if there are no dates
If one or more officeholders have a P39 with no dates, they will be
displayed at the end of the table. But because they are notionally still
positioned in the table, the earliest officeholder will try to compare
dates with the notional predecessor. In such a case we don't want to
complain about the dates don't seem to sync (or are incomparable).