Set dummy credentials for tests
The test require _some_ credentials to be set, but we want these to be
consistent (rather than picked up from the environment) and not be real
ones. The tests shouldn't actually be _using_ these in any way (and it's
probably a warning sign that an error is raised until just before the
code actually tries to make a remote call using them), but for now we
should at least be explicit about them.