Improve confusing replaces/replaced-by warnings
I always find the text for inconsistent succession a little bit
confusing. Rather than simply saying that one value differs from
another, which requires a little bit of effort to decipher, make it a
little clearer that the qualifier on the P39 is being compared to the
following (or preceding) value in this table.