Configure RSpec cops
minitest doesn't have the #described_class method, and seems unlikely to
add it (https://github.com/seattlerb/minitest/issues/526) so
RSpec/DescribedClass is meaningless in this context.

We should probably only allow one expectation per spec, but fixing those
is a quite a lot of work, so let's allow 2 for now.