Have different configuration classes for position types
Rather than cluttering up the Report class with lots of "if this type of
position, then this, else this" logic, factor the different
configuration data out into Config objects.

It's not entirely clear yet what other data will vary, so keep the
interface to these as lightweight as possible for now.